# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Postaway",
    "description": "This is the API for a simple social media platform."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Localhost server"
    }
  ],
  "components": {
    "securitySchemes": {
      "JWT": {
        "in": "header",
        "name": "Authorization",
        "type": "apiKey"
      }
    }
  },
  "paths": {
...
    "/api/post": {
      "get": {
        "tags": ["Post"],
        "summary": "Get posts",
        "description": "The user can retrieve paginated posts by giving the page number. The user can also provide sort order as 'asc' or 'desc'.",
        "security": [
          {
            "JWT": {}
          }
        ],
        "parameters": [
          {
            "name": "pageNo",
            "in": "query",
            "description": "Page number",
            "required": false,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "Sort order - 'asc' or 'desc'",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Posts retrieved successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "posts": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "integer"
                          },
                          "userId": {
                            "type": "integer"
                          },
                          "caption": {
                            "type": "string"
                          },
                          "imageUrl": {
                            "type": "string"
                          },
                          "status": {
                            "type": "string"
                          },
                          "createdAt": {
                            "type": "string"
                          },
                          "updatedAt": {
                            "type": "string"
                          },
                          "likes": {
                            "type": "integer"
                          },
                          "comments": {
                            "type": "integer"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "No posts found."
          }
        }
      },
      "post": {
        "tags": ["Post"],
        "summary": "Create a post",
        "description": "The user can create a new post by providing a caption and a file.",
        "security": [
          {
            "JWT": {}
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "caption": {
                    "type": "string"
                  },
                  "image": {
                    "type": "string",
                    "format": "binary"
                  }
                },
                "required": ["caption", "image"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Post created successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    },
                    "post": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "integer"
                        },
                        "userId": {
                          "type": "integer"
                        },
                        "caption": {
                          "type": "string"
                        },
                        "imageUrl": {
                          "type": "string"
                        },
                        "status": {
                          "type": "string"
                        },
                        "createdAt": {
                          "type": "string"
                        },
                        "updatedAt": {
                          "type": "string"
                        },
                        "likes": {
                          "type": "integer"
                        },
                        "comments": {
                          "type": "integer"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}